+++ src/LaravelDDD.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Lunarstorm\LaravelDDD;
-
-class LaravelDDD
-{
-}