+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to `laravel-ddd` will be documented in this file.
 
 ## [Unversioned]
+### Added
+- Ability to ignore folders during autoloading via config(`ddd.autoload.ignore`), or register a custom filter callback via `DDD::filterAutoloadPathsUsing(callable $filter)`.
+
 ### Changed
 - Internal: Domain cache is no longer quietly cleared on laravel's `cache:clearing` event, so that `ddd:cache` yields consistent results no matter what order runs in production (before or after `cache:clear` or `optimize:clear` commands).